Fix postal_code for batch geocodio geocoding
Fixes #178 by renaming parameters to match API specifications.

I followed the general framework used in query_factory.R to make it robust to any other differences in parameter names, but the same result could be accomplished by hard coding the rename into the batch_geocodio() function as it is done in some of the other batch functions.